requirements.txt
# Requirements for deployment
streamlit
google-generativeai
python-dotenv
